Rename yomichan-util.js to util.js
