+++ dev/build.js
@@ -19,7 +19,7 @@ const fs = require('fs');
 const path = require('path');
 const readline = require('readline');
 const childProcess = require('child_process');
-const util = require('./yomichan-util');
+const util = require('./util');
 const {getAllFiles, getDefaultManifestAndVariants, createManifestString} = util;
